Bugfix with RSS
Added checking for the tables existence.
Without this, sometimes there were errors, for example, when blogs are
not installed but user try to read blogs RSS.

This fix allows to prevent errors with RSS.

+++ infusions/rss_feeds_panel/feeds/rss_downloads.php
@@ -7,6 +7,7 @@
 | Filename: rss_downloads.php
 | Author: Robert Gaudyn (Wooya)
 | Co-Author: Joakim Falk (Domi)
+| Co-Author: Chubatyj Vitalij (Rizado)
 +--------------------------------------------------------+
 | This program is released as free software under the
 | Affero GPL license. You can redistribute it and/or
@@ -26,33 +27,36 @@ if (file_exists(INFUSIONS."rss_feeds_panel/locale/".LANGUAGE.".php")) {
 	include INFUSIONS."rss_feeds_panel/locale/English.php";
 }
 
-$result = dbquery("SELECT tbl1.*, tbl2.* FROM ".DB_DOWNLOAD_CATS." tbl1
-RIGHT JOIN ".DB_DOWNLOADS." tbl2 ON tbl1.download_cat_id=tbl2.download_cat
-WHERE ".groupaccess('download_cat_access').(multilang_table("DL")?" AND download_cat_language='".LANGUAGE."'":"")."
-ORDER BY tbl2.download_count DESC LIMIT 0,10");
+if (db_exists(DB_DOWNLOADS) && db_exists(DB_DOWNLOAD_CATS)) {
+
+	$result = dbquery("SELECT tbl1.*, tbl2.* FROM ".DB_DOWNLOAD_CATS." tbl1
+	RIGHT JOIN ".DB_DOWNLOADS." tbl2 ON tbl1.download_cat_id=tbl2.download_cat
+	WHERE ".groupaccess('download_cat_access').(multilang_table("DL")?" AND download_cat_language='".LANGUAGE."'":"")."
+	ORDER BY tbl2.download_count DESC LIMIT 0,10");
 
 	echo "<?xml version=\"1.0\" encoding=\"".$locale['charset']."\"?>\n\n";
 	echo "<rss version=\"2.0\">\n\n <channel>\n";
 
-if (dbrows($result) != 0) {
+	if (dbrows($result) != 0) {
 
-echo "<title>".$settings['sitename'].$locale['rss003'].(multilang_table("DL")?" ".$locale['rss007']." ".LANGUAGE:"")."</title>\n<link>".$settings['siteurl']."</link>\n";
-echo "<description>".$settings['description']."</description>\n";
+		echo "<title>".$settings['sitename'].$locale['rss003'].(multilang_table("DL")?" ".$locale['rss007']." ".LANGUAGE:"")."</title>\n<link>".$settings['siteurl']."</link>\n";
+		echo "<description>".$settings['description']."</description>\n";
 
-while ($row=dbarray($result)) {
-	$rsid = intval($row['download_id']);
-	$rtitle = $row['download_title'];
-	$description = stripslashes(nl2br($row['download_description']));
-	$description = strip_tags($description, "<a><p><br /><br /><hr />");
-	echo "<item>\n<title>".htmlspecialchars($rtitle)."</title>\n";
-	echo "<link>".$settings['siteurl']."downloads.php?download_id=".$rsid."</link>\n";
-	echo "<description>".htmlspecialchars($description)."</description>\n";
-	echo "</item>\n\n";
+		while ($row=dbarray($result)) {
+			$rsid = intval($row['download_id']);
+			$rtitle = $row['download_title'];
+			$description = stripslashes(nl2br($row['download_description']));
+			$description = strip_tags($description, "<a><p><br /><br /><hr />");
+			echo "<item>\n<title>".htmlspecialchars($rtitle)."</title>\n";
+			echo "<link>".$settings['siteurl']."downloads.php?download_id=".$rsid."</link>\n";
+			echo "<description>".htmlspecialchars($description)."</description>\n";
+			echo "</item>\n\n";
+		}
+	} else {
+		echo "<title>".$settings['sitename'].$locale['rss004']."</title>\n
+		<link>".$settings['siteurl']."</link>\n
+		<description>".$locale['rss008']."</description>\n";
+	}
+	echo "</channel></rss>";
 }
-} else {
-	echo "<title>".$settings['sitename'].$locale['rss004']."</title>\n
-	<link>".$settings['siteurl']."</link>\n
-	<description>".$locale['rss008']."</description>\n";
-}
-echo "</channel></rss>";
 ?>
